basic implementation of ignite done

+++ inc/metadata.h
@@ -1,8 +1,8 @@
 struct metadata {
     uint8_t entry_format;
     uint8_t branch_type;
-    uint8_t branch_delta;
-    uint8_t target_delta;
+    int branch_delta;
+    int target_delta;
     uint64_t full_addr;
 
     metadata() : entry_format(0), branch_type(0), branch_delta(0), target_delta(0) {}
